+++ android/app/src/main/kotlin/com/healtine/workout_player/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.healtine.workout_player
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}