+++ lib/view/widgets/miniplayer/miniplayer_title.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:workout_player/models/routine.dart';
-import 'package:workout_player/models/youtube_video.dart';
+// import 'package:workout_player/models/youtube_video.dart';
 import 'package:workout_player/styles/text_styles.dart';
 import 'package:workout_player/utils/formatter.dart';
 import 'package:workout_player/view_models/miniplayer_model.dart';
-import 'package:youtube_plyr_iframe/youtube_plyr_iframe.dart';
-import 'package:collection/collection.dart';
+// import 'package:youtube_plyr_iframe/youtube_plyr_iframe.dart';
+// import 'package:collection/collection.dart';
 
 class MiniplayerTitle extends StatelessWidget {
   const MiniplayerTitle({
@@ -52,28 +52,29 @@ class MiniplayerTitle extends StatelessWidget {
         );
       }
     } else {
-      return YoutubeValueBuilder(
-        builder: (context, value) {
-          final video = model.currentWorkout as YoutubeVideo?;
-          final currentWorkout = video!.workouts
-                  .lastWhereOrNull((e) => e.position <= value.position) ??
-              video.workouts[0];
-          int currentIndex = video.workouts
-              .lastIndexWhere((e) => e.position <= value.position);
+      return Container();
+      // return YoutubeValueBuilder(
+      //   builder: (context, value) {
+      //     final video = model.currentWorkout as YoutubeVideo?;
+      //     final currentWorkout = video!.workouts
+      //             .lastWhereOrNull((e) => e.position <= value.position) ??
+      //         video.workouts[0];
+      //     int currentIndex = video.workouts
+      //         .lastIndexWhere((e) => e.position <= value.position);
 
-          currentIndex = (currentIndex == -1) ? 0 : currentIndex;
+      //     currentIndex = (currentIndex == -1) ? 0 : currentIndex;
 
-          model.updateCurrentIndex(currentIndex);
+      //     model.updateCurrentIndex(currentIndex);
 
-          return Text(
-            Formatter.localizedTitle(
-              currentWorkout.workoutTitle,
-              currentWorkout.translatedWorkoutTitle,
-            ),
-            style: textStyle,
-          );
-        },
-      );
+      //     return Text(
+      //       Formatter.localizedTitle(
+      //         currentWorkout.workoutTitle,
+      //         currentWorkout.translatedWorkoutTitle,
+      //       ),
+      //       style: textStyle,
+      //     );
+      //   },
+      // );
     }
   }
 }