+++ lib/view/widgets/miniplayer/youtube_player_widget.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter/gestures.dart';
+// import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:workout_player/view_models/main_model.dart';
-import 'package:workout_player/view_models/miniplayer_model.dart';
-import 'package:youtube_plyr_iframe/youtube_plyr_iframe.dart';
+// import 'package:workout_player/view_models/miniplayer_model.dart';
+// import 'package:youtube_plyr_iframe/youtube_plyr_iframe.dart';
 
 class YoutubePlayerWidget extends ConsumerWidget {
   const YoutubePlayerWidget({
@@ -15,16 +15,16 @@ class YoutubePlayerWidget extends ConsumerWidget {
   Widget build(BuildContext context, ScopedReader watch) {
     logger.d('[YoutubePlayerWidget] building');
 
-    final model = watch(miniplayerModelProvider);
+    // final model = watch(miniplayerModelProvider);
     final size = MediaQuery.of(context).size;
 
     return SizedBox(
       height: 220,
       width: size.width,
-      child: YoutubePlayerIFrame(
-        gestureRecognizers: const <Factory<OneSequenceGestureRecognizer>>{},
-        controller: model.youtubeController,
-      ),
+      // child: YoutubePlayerIFrame(
+      //   gestureRecognizers: const <Factory<OneSequenceGestureRecognizer>>{},
+      //   controller: model.youtubeController,
+      // ),
     );
   }
 }