+++ lib/view_models/settings_tab_model.dart
@@ -68,7 +68,7 @@ class SettingsTabModel with ChangeNotifier {
     showAboutDialog(
       context: context,
       applicationName: S.current.applicationName,
-      applicationVersion: 'v.0.3.6',
+      applicationVersion: 'v.0.3.6+1',
       applicationIcon: SvgPicture.asset(
         'assets/svgs/break_logo.svg',
         width: 36,